+++ mycodo/inputs/atlas_ec.py
@@ -300,7 +300,7 @@ class InputModule(AbstractInput):
                     break
 
         elif self.interface == 'I2C':
-            return_string = atlas_return
+            return_string = self.atlas_device.build_string(atlas_return)
 
         if return_string and ',' in return_string:
             # Multiple values returned